requirements.txt
streamlit==1.33.0
requests==2.31.0
matplotlib==3.8.4
python-dotenv